+++ src/wallet/infrastructure/models.py
@@ -1,7 +1,7 @@
 import uuid
 
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy import UUID, FLOAT
+from sqlalchemy import UUID, DECIMAL
 
 from WalletApp.src.core.db import Base
 
@@ -10,4 +10,4 @@ class WalletModel(Base):
     __tablename__ = "wallets"
 
     id: Mapped[uuid.UUID] = mapped_column(UUID, primary_key=True, default=uuid.uuid4())
-    balance: Mapped[float] = mapped_column(FLOAT, nullable=False)
+    balance: Mapped[float] = mapped_column(DECIMAL, nullable=False)
